# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Localfluence API - influencer marketplace
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  business: "business",
  influencer: "influencer",
} as const;

export interface SignupBody {
  email: string;
  /** @minLength 6 */
  password: string;
  role: Role;
  name: string;
  city?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface CurrentUser {
  id: number;
  email: string;
  role: Role;
  name?: string;
  avatarUrl?: string;
}

export interface AuthResponse {
  token: string;
  user: CurrentUser;
}

export interface Profile {
  userId: number;
  role: Role;
  name?: string;
  bio?: string;
  city?: string;
  area?: string;
  category?: string;
  instagramUrl?: string;
  youtubeUrl?: string;
  followers?: number;
  avatarUrl?: string;
  coverUrl?: string;
  companyName?: string;
  industry?: string;
  description?: string;
}

export interface UpdateProfileBody {
  name?: string;
  bio?: string;
  city?: string;
  area?: string;
  category?: string;
  instagramUrl?: string;
  youtubeUrl?: string;
  followers?: number;
  avatarUrl?: string;
  coverUrl?: string;
  companyName?: string;
  industry?: string;
  description?: string;
}

export interface InfluencerSummary {
  id: number;
  name: string;
  city: string;
  area?: string;
  category: string;
  followers: number;
  avatarUrl?: string;
  bio?: string;
}

export type InfluencerDetail = InfluencerSummary & {
  instagramUrl?: string;
  youtubeUrl?: string;
  coverUrl?: string;
};

export interface CategoryCount {
  category: string;
  count: number;
}

export interface CityCount {
  city: string;
  count: number;
}

export interface Campaign {
  id: number;
  businessId: number;
  businessName?: string;
  title: string;
  description: string;
  budget: number;
  location: string;
  deliverables: string;
  createdAt: string;
}

export interface CreateCampaignBody {
  title: string;
  description: string;
  budget: number;
  location: string;
  deliverables: string;
}

export type RequestStatus = (typeof RequestStatus)[keyof typeof RequestStatus];

export const RequestStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
} as const;

export interface CollaborationRequest {
  id: number;
  businessId: number;
  businessName?: string;
  influencerId: number;
  influencerName?: string;
  campaignId?: number;
  campaignTitle?: string;
  message?: string;
  status: RequestStatus;
  createdAt: string;
}

export interface CreateRequestBody {
  influencerId: number;
  campaignId?: number;
  message?: string;
}

export interface UpdateRequestStatusBody {
  status: RequestStatus;
}

export interface Message {
  id: number;
  senderId: number;
  receiverId: number;
  content: string;
  createdAt: string;
}

export interface SendMessageBody {
  /** @minLength 1 */
  content: string;
}

export interface Conversation {
  otherUserId: number;
  otherUserName: string;
  otherUserRole?: Role;
  otherUserAvatarUrl?: string;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
}

export interface BusinessDashboard {
  campaignCount: number;
  requestsSent: number;
  activeCollaborations: number;
  pendingRequests?: number;
  recentRequests: CollaborationRequest[];
  recentCampaigns: Campaign[];
}

export interface InfluencerDashboard {
  pendingRequests: number;
  acceptedCollaborations: number;
  totalRequests: number;
  recentRequests: CollaborationRequest[];
}

export type ListInfluencersParams = {
  city?: string;
  area?: string;
  category?: string;
  minFollowers?: number;
  maxFollowers?: number;
  q?: string;
};

export type ListMyRequestsParams = {
  status?: ListMyRequestsStatus;
};

export type ListMyRequestsStatus =
  (typeof ListMyRequestsStatus)[keyof typeof ListMyRequestsStatus];

export const ListMyRequestsStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
} as const;