# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Localfluence API - influencer marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupBody" }
      responses:
        "200":
          description: Signed up
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
...
  /profiles/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      responses:
        "200":
          description: My profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
    put:
      operationId: updateMyProfile
      tags: [profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileBody" }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
  /influencers:
    get:
      operationId: listInfluencers
      tags: [influencers]
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: area
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: minFollowers
          in: query
          schema: { type: integer }
        - name: maxFollowers
          in: query
          schema: { type: integer }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Influencer list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InfluencerSummary" }
...
  /influencers/{id}:
    get:
      operationId: getInfluencer
      tags: [influencers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Influencer detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InfluencerDetail" }
  /campaigns:
    get:
      operationId: listMyCampaigns
      tags: [campaigns]
      responses:
        "200":
          description: Campaigns owned by current business
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Campaign" }
    post:
      operationId: createCampaign
      tags: [campaigns]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCampaignBody" }
      responses:
        "200":
          description: Created campaign
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
...
  /messages/with/{userId}:
    get:
      operationId: listMessagesWithUser
      tags: [messages]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Message" }
    post:
      operationId: sendMessageToUser
      tags: [messages]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageBody" }
      responses:
        "200":
          description: Sent message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
...
components:
  schemas:
...
    Role:
      type: string
      enum: [business, influencer]
    SignupBody:
      type: object
      required: [email, password, role, name]
      properties:
        email: { type: string }
        password: { type: string, minLength: 6 }
        role: { $ref: "#/components/schemas/Role" }
        name: { type: string }
        city: { type: string }
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/CurrentUser" }
    CurrentUser:
      type: object
      required: [id, email, role]
      properties:
        id: { type: integer }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        name: { type: string }
        avatarUrl: { type: string }
    Profile:
      type: object
      required: [userId, role]
      properties:
        userId: { type: integer }
        role: { $ref: "#/components/schemas/Role" }
        name: { type: string }
        bio: { type: string }
        city: { type: string }
        area: { type: string }
        category: { type: string }
        instagramUrl: { type: string }
        youtubeUrl: { type: string }
        followers: { type: integer }
        avatarUrl: { type: string }
        coverUrl: { type: string }
        companyName: { type: string }
        industry: { type: string }
        description: { type: string }
    UpdateProfileBody:
      type: object
      properties:
        name: { type: string }
        bio: { type: string }
        city: { type: string }
        area: { type: string }
        category: { type: string }
        instagramUrl: { type: string }
        youtubeUrl: { type: string }
        followers: { type: integer }
        avatarUrl: { type: string }
        coverUrl: { type: string }
        companyName: { type: string }
        industry: { type: string }
        description: { type: string }
    InfluencerSummary:
      type: object
      required: [id, name, city, category, followers]
      properties:
        id: { type: integer }
        name: { type: string }
        city: { type: string }
        area: { type: string }
        category: { type: string }
        followers: { type: integer }
        avatarUrl: { type: string }
        bio: { type: string }
    InfluencerDetail:
      allOf:
        - $ref: "#/components/schemas/InfluencerSummary"
        - type: object
          properties:
            instagramUrl: { type: string }
            youtubeUrl: { type: string }
            coverUrl: { type: string }
...
    Campaign:
      type: object
      required: [id, businessId, title, description, budget, location, deliverables, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        businessName: { type: string }
        title: { type: string }
        description: { type: string }
        budget: { type: integer }
        location: { type: string }
        deliverables: { type: string }
        createdAt: { type: string, format: date-time }
    CreateCampaignBody:
      type: object
      required: [title, description, budget, location, deliverables]
      properties:
        title: { type: string }
        description: { type: string }
        budget: { type: integer }
        location: { type: string }
        deliverables: { type: string }
...
    Message:
      type: object
      required: [id, senderId, receiverId, content, createdAt]
      properties:
        id: { type: integer }
        senderId: { type: integer }
        receiverId: { type: integer }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    SendMessageBody:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
...