# lib/api-zod/src/generated/types/influencerDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Localfluence API - influencer marketplace
 * OpenAPI spec version: 0.1.0
 */
import type { InfluencerSummary } from "./influencerSummary";

export type InfluencerDetail = InfluencerSummary & {
  instagramUrl?: string;
  youtubeUrl?: string;
  coverUrl?: string;
};