# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Localfluence API - influencer marketplace
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface Profile {
  userId: number;
  role: Role;
  name?: string;
  bio?: string;
  city?: string;
  area?: string;
  category?: string;
  instagramUrl?: string;
  youtubeUrl?: string;
  followers?: number;
  avatarUrl?: string;
  coverUrl?: string;
  companyName?: string;
  industry?: string;
  description?: string;
}