# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Localfluence API - influencer marketplace
 * OpenAPI spec version: 0.1.0
 */

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  business: "business",
  influencer: "influencer",
} as const;