# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Localfluence API - influencer marketplace
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface SignupBody {
  email: string;
  /** @minLength 6 */
  password: string;
  role: Role;
  name: string;
  city?: string;
}